Adds functions to templates
- Functions support callback such that more complex operations can be done as part of templates
- Also some minor updates to the readme
