+++ lib/jyson.js
@@ -16,7 +16,7 @@ const getKey = (object, key) => {
   });
 };
 
-const fillKeys = (json, object) => {
+const fillKeys = (json, object, opts) => {
   const jsonKeys = Object.keys(json);
 
   _.each(jsonKeys, key => {
@@ -38,9 +38,13 @@ const fillKeys = (json, object) => {
       return json[key] = getKey(object, json[key][0], null);
     }
 
+    if (_.isFunction(json[key])) {
+      return json[key] = json[key]({object, key, opts});
+    }
+
     // Nested Object
     if(_.isObject(json[key])) {
-      return json[key] = fillKeys(json[key], object);
+      return json[key] = fillKeys(json[key], object, opts);
     }
 
     // We should not be able to get here
@@ -51,16 +55,16 @@ const fillKeys = (json, object) => {
 };
 
 module.exports.buildTemplateFunction = (templateObject) => {
-  const templateFunction = (objects) => {
+  const templateFunction = (objects, opts = {}) => {
 
     if (!(objects instanceof Array)) {
-      return templateFunction([objects])[0];
+      return templateFunction([objects], opts)[0];
     }
 
     return objects.map(object => {
       const json = _.cloneDeep(templateObject);
 
-      fillKeys(json, object);
+      fillKeys(json, object, opts);
 
       return json;
     });
