+++ package.json
@@ -25,7 +25,9 @@
     "chai": "^4.0.0",
     "eslint": "^3.19.0",
     "mocha": "^3.4.2",
-    "q": "^1.5.0"
+    "q": "^1.5.0",
+    "sinon": "^2.3.2",
+    "sinon-chai": "^2.10.0"
   },
   "dependencies": {
     "lodash": "^4.17.4"
